+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.3.2] - 2019-11-07
+### Updated
+- PHPUnit version.
+- WebDriver version.
+### Fixed
+- Check Settings value transfer.
+- Tests - WebDriver now quits even if EyesWebDriver failed to create.
+- MatchWindowTask.
+- Images tests.
+### Added
+- Eyes->getLogger() method.
+
 ## [1.3.1] - 2019-10-14
 ### Added 
 - This CHANGELOG file.